Rename get-advice directory
